+++ DTOs/PendingUserDto.cs
@@ -5,13 +5,13 @@ namespace UABackbone_Backend.DTOs;
 public class PendingUserDto
 {
     [Required]
-    public required ushort Id { get; set; }
+    public required int      Id          { get; set; }
     [Required]
-    public required string Username { get; set; } = null!;
+    public required string   Username    { get; set; } = null!;
     [Required]
-    public required string Email { get; set; } = null!;
+    public required string   Email       { get; set; } = null!;
     [Required]
-    public required short Local { get; set; }
+    public required int      Local       { get; set; }
     [Required]
     public required DateTime SubmittedAt { get; init; }
 }
