+++ DTOs/UserDto.cs
@@ -5,16 +5,16 @@ namespace UABackbone_Backend.DTOs;
 public class UserDto
 {
     [Required] 
-    public required ushort Id { get; set; }
+    public required int    Id        { get; set; }
     [Required]    
-    public required string Username { get; set; } = null!;
+    public required string Username  { get; set; } = null!;
     [Required]
     public required string FirstName { get; set; } = null!;
     [Required]
-    public required string LastName { get; set; } = null!;
+    public required string LastName  { get; set; } = null!;
     [Required]
-    public required string Email { get; set; } = null!;
+    public required string Email     { get; set; } = null!;
     [Required]
-    public required short Local { get; set; }
-    public string Token { get; set; } = null!;
+    public required int    Local     { get; set; }
+    public string          Token     { get; set; } = null!;
 }
